Patch for WL#5639 "Enable Index Condition Pushdown optimizer feature"
With this change the MySQL server will be built with support for the 
Index Condition Pushdown (ICP) feature. The default will be that ICP
will be enabled.

The following changes to MySQL are implemented:

1. Enable index condition pushdown
  
With this feature enabled the optimizer will consider to push down to the
handler/storage engine parts (or all) of the where condition that can be
evaluated by the handler/storage engine using the content of the index that is
used for execution of the query. 

The main change by enabling index condition pushdown should be improved
performance for certain queries. The user should be able to see if index
condition pushdown has been used by looking in the "Extra" column of the EXPLAIN
output. If this field contains the string "Using index condition" parts or all
of the where statement has been pushed down and will be evaluated by the
handler/storage engine.

2. Enable the user to control the use of index condition pushdown

When this has been enabled the user will be able to control the use 
of index condition pushdown by using the optimizer_switch configuration variable. 
The new optimizer_switch flag that will become available is named
"index_condition_pushdown" and can take the values "on", "off" and "default".
The default value for this optimizer switch flag will be "on".