Fix Bug #57854: assert table->n_rec_locks == 0 in DROP TABLE after recovery
In lock_remove_all_on_table() we iterate over table->locks to delete all the locks on
that table held by transactions.  From table->locks we get the list of transactions that
have record locks on the that table. For recovered transactions that wont work because they
can have implicit record locks converted to explicit record locks without any covering
table level IS or IX locks.
			            
The fix is to handle the case for such recovered transactions. Also, this is an issue
in pre dict LRU code too but there we don't have stringent checks.

rb://512, Approved by Jimmy Yang.