BUG#58063: rpl_server_uuid segfaults on win
While connecting to the master the slave thread sets a session
variable (@master_binlog_checksum). Then, later, it queries the
master for this variable content. However, in this test, the
master is restarted a few times, which can cause the slave to
reconnect between the time it sets and the time it queries the 
session var @master_binlog_checksum on the master. 

Ultimately, this can cause the slave to receive NULL when
SELECTing the var after a master restart, causing the slave to 
crash.

We fix this by checking whether the value is NULL or not before
taking any action based on the var value.

Other changes in this cset:

1. Fix for rpl_checksum_cache failure in valgrind

   The test takes too long when using RBR and times out because
   the slave takes too long time to replay the row events for a
   table without a primary key. We fix it by adding a PK to the
   table, which makes the slave to find the rows to delete.

2. rpl_corruption fails with errors in valgrind

   This is a bug uncovered while testing WL#2540. Details are
   available in BUG#58129. We workaround it by replacing the
   offending part of the test (ie, do not use TEXT type and use
   VARCHAR instead).