Fix for Bug#56529 Crash due to long semaphore wait in InnoDB with ICP and subqueries
This crash was caused by the server pushing down an index condition to
InnoDB that contained a subselect. When InnoDB executed this index
condition the subselect would open a new handler object and access
InnoDB a second time. This caused the same transaction to have two
active data accesses to InnoDB which triggered the assert in InnoDB.

The fix for this is to not push down index conditions to the handler
that contain a subselect.