Bug#54641: semijoin loosescan gives duplicate rows
The loosescan semijoin strategy does not handle join buffering 
on any of the tables in the loosescan range. Prior to this bug
fix, check_join_cache_usage() would not do join buffering on a 
join table if the loosescan_match_tab pointer was set for it. 

However, loosescan_match_tab is a pointer from the first to the 
last join table handled by the loosescan strategy. Join 
buffering could therefore be used on all join tables except 
the first in the loosescan range.

The fix is to skip join buffering if the join table is in 
the loosescan range, i.e., if
join_tab->get_sj_strategy() == SJ_OPT_LOOSE_SCAN,
instead of checking loosescan_match_tab.