Bug #58411 :wixca project fails to build when using Express Edition of VS:
The problem was MFC header file includes into the resource definition file (CustomAction.rc)
afxres.h is not available with VS Express.

The fix is to remove resource file from compilation and souce code repository. version of
custom action dll is of no interest for anyone, it is internal dll kept inside the MSI.