Bug#47699 rpl.rpl_backup_block fails sporadically
It failed because an error appeared in Slave_SQL_Error,
the test case expects that no error happens. The reason is
that START SLAVE released the lock and returned before it cleared
the error, so there was a possibility that Slave_SQL_Error was not
0 when doing the following:
START SLAVE SQL_THREAD;
SHOW SLAVE STATUS;
      
After this patch, clearing error always happens before START SLAVE
releases the lock.