Bug #57035 'ulonglong sql_mode' stored into ulong, is risky            and causes compiler warning
In 5.5 THD::variables::sql_mode was changed from ulong to
ulonglong. The problem was that several variables used to
store sql_mode were still declared as ulong. This patch
updates these declarations to sql_mode_t which is typedef'ed
to ulonglong.

The patch also adds a test case designed to break once
someone adds a new SQL mode. This is done to serve as an
indication that current tests which use
"SET @@sql_mode= cast(pow(2,32)-1" to activate all SQL modes
will then have to be updated.