WL#5692 Enable Disk-Sweep Multi-Range Read optimizer feature.
This patch contains the first part of this worklog. It changes the default
value of the optimizer switch "mrr_cost_based" from being off as default
to be on by default.

To ensure that both non-cost-based MRR and cost-based MRR get tested the
following changes are done to our MRR tests:

1. to ensure MRR code get tested: the optimizer_switch flag mrr_cost_based
   is set to off in test that specifically is named to test MRR (these
   have _mrr or _all in their name).

2. to ensure cost-based MRR get tested: New versions of existing MRR tests
   are added. These have added _mrr_cost in their test name.