Bug#58620 Performance schema failed assert, pfs_instr_class.cc, line 956
Before this fix, the member PFS_table_share::m_refcount was maintained
with ++ and -- operations.

This is not thread safe, since two threads may open or close table
handles on the same table share, and compete when updating m_refcount.

This fix uses atomic operations to maintain this reference counter,
making the counter safe.