Bug#56380: valgrind memory leak warning from partition tests
There could be memory leaks if ALTER ... PARTITION command fails.

Problem was that the list of items to free was not set in
the partition info structure when fix_partition_func call failed
during ALTER ... PARTITION.

Solved by always setting the list in the partition info struct.

sql/table.cc:
  item_free_list is not set if (!work_part_info_used) and
  fix_partition_func failed. Which may result in a mem leak.