Bug#58172 Column performance_schema.threads.processlist_time out of range
Before this fix, the column processlist_time in table performance_schema.threads
could receive out of range values.

Some negative time was computed, because timers may have some drift and are not
guaranteed to be always monotonous.

This fix replaces negative time values with 0, to correct the data,
and make the data fit into an unsigned column.