BUG#58762: rpl.rpl_packet fails in the weekly PB2 run
The test contains an assertion that depends on an exact
match on mysqlbinlog output. However, mysqlbinlog output
changes when checksums are ON, because it adds the 
checksum description as well. This makes the test case to
fail when running it with --mysqld=--binlog-checksum=...

The fix is to rewrite the assertion so that the pattern
matching is made crc agnostic. To do this we resort to
perl instead of mysql text function INSTR.