BuG#58913: rpl_lcase_tblnames_rewrite_db & rpl_stm_lcase_tblnames            fails on Mac OSX
The following three test cases fail on Mac OSX because they 
assume that the lower_case_table_names is set to 0 on the master
by default (which is not the case on MacOSX - defaults to 2 - 
nor on Windows - defaults to 1): 

  - rpl.rpl_row_lcase_tblnames
  - rpl.rpl_stm_lcase_tblnames
  - rpl.rpl_lcase_tblnames_rewrite_db

This causes a result mismatch when the tests are run on such
platforms. However, the tests are configured to be skipped on
windows, so this issue does not show up on that platform, only 
on Mac OSX.

We fix this by masking out the part of the result that is 
different on Mac OSX and other platforms.