Bug#21437 server_errno=29 error message flood mysqld error log
If an error happens while dumping a binary log from the master,
the master sends one of the specific set of error messages and
the slave I/O thread stops immediately.
            
However, when a slave requests a binlog file which does not exist
on master, the master sends 'EE_FILENOTFOUND' error (i.e. error code 29)
to the slave. the 'EE_FILENOTFOUND' is a low level error and slave
I/O thread will still retry to request the binlog file, thus is flooding
the mysqld error log.
          
This happens because the master just sends the first error message set in 
the diagnostic area and as such the I/O thread does not stop. To fix the
the problem, we use a temporary diagnostic area to store the low level errors.
MASTER_FATAL_ERROR_READING_BINLOG is always set into dump thread's real
diagnostic area and sent to slave I/O thread.