WL#5493 Binlog crash-safe when master crashed
Trim the crashed binlog file to last valid transaction or event
(non-transaction) base on binlog size(valid_pos) when MYSQL server
crashed in the middle of binlog. And add a temp file to guarantee
the binlog index file to be crash safe.

WL#5440 Test binlog and replication when master crashed

Test whether the transaction is binlogged partly and the replication
works fine or not when the master crashed.