BUG#59079 rpl_binlog_errors.test fails when enables binlog-checksum
The test case dependent on the size of incident event to calculate
the event offset, howerver, the size of Incident_log_event is different when binlog-checksum
is enabled, and caused the test case to fail.

Fixed the problem by use the event sequence number instead event
log file position to locate the incident event.