Patch for Bug#58463 Error Can't find record on SELECT with JOIN and ORDER BY
This problem was caused by the server pushing down an index condition on
the primary key as part of the optimization phase. Later, during the 
execution phase the server (due to a bug) changed to use a secondary index
instead of the primary index for accessing the table. This made MRR fail
in phase 3 due to not finding the expected records when retrieving them
using the primary key due to the pushed index condition filtering them away.

This problem with the server accidentally switching to using a secondary index will
be fixed by the fix for Bug#58456. 

This patch contains a test case based on the bug report and an added assert to
the DS-MRR implementation that will trigger if a similar situation should occur
again (DS-MRR used on a secondary index while an index condition has been pushed
on the primary index for the same handler object).