BUG#59123 rpl_stm_binlog_max_cache_size fails sporadically with found warnings
rpl_stm_binlog_max_cache_size.test fails sporadically on PB2 
daily-trunk-bugfixing with binlog-checksum enabled, the expected
incident error message was not properly suppressed, and the
incident message would have extra strange characters appended.

The cause is that the incident error message restored from log event
is not properly ended with '\0'. This does not cause problem normally
because the error message is at the end of the event, and we add
a '\0' at the end of an event. but when binlog-checksum is enabled,
the checksum value will be after the message, and so extra strange
characters could be shown and cause the suppression rule to fail.

Fixed the problem by allocate memory for the string and added '\0'
at the end of the message.