Bug#58918: Add a clone member function to the Field class for better duplication
The patch removes all copying of Field objects using memcpy() and replaces it
with virtual clone() and clone(MEM_ROOT*) member functions that rely on the
copy constructor. An operator new (MEM_ROOT*) is also implemented. The sole
purpose of the Field::size_of() member function was to support the memcpy and
it is now removed.

The previously existing clone member function has been inlined in the only place
where it was called from.