Fix for Bug#58816 Extra temporary duplicate rows in result set when                   switching ICP off.
The wrong result was caused by that the handler object was in an
inconsistent state when running the query. The same handler object
was previously used by an explain for the same query. 

When doing the explain the optimizer pushed down an index condition to
the handler. This set information about the pushed index condition
as well as setting the member variable in_range_check_pushed_down to
true. The inconsistency that this resulted in was that when the
explain statement completed and the handler object was ready for
re-use the information about the pushed index condition was reset but
the in_range_check_pushed_down was still true.

When executing the same query (after disabling index condition
pushdown) using the same handler object with the
in_range_check_pushed_down still being true caused that neither the
server nor the storage engine were performing the range check. This
resulted in that we would read to the end of the table for the first
range instead of stopping at the end of range criterion.

The fix for this problem is to add code to handler::ha_reset() to
to reset both information about the pushed index condition and the
in_range_check_pushed_down. This method is called when a
statement closes its tables and the handler object is made ready
for reuse.