Fix for bug#58628: Incorrect result for 'WHERE NULL NOT IN (<subquery>)
create_ref_for_key() allowed any constant part of a REF key to be evaluated
and stored into the 'key_buff' during ::optimize(). These 'store_key*' was
*not* kept in ref.key_copy[] as they where constant and we assumed we
would not have to reevaluate them during JOIN::exec()
      
However, during execute NULL values in REF key has to be detected as they may
need special attention - as in 'Full scan on NULL key'. This is done by 
subselect_uniquesubquery_engine::copy_ref_key() which check if any keyparts 
evaluated to a NULL-value.
      
As we didn't keep a store_key for a constant value, a NULL-constant was not
detected by subselect_uniquesubquery_engine::copy_ref_key() !
      
This fixs modifies create_ref_for_key() to check if a NULL-value constant 
was produced - In these cases it keeps the store_key, which then will be 
reevaluated in JOIN::exec() and trigger correct handling of NULL-valued keys.